refactoring of render and grouping, widget spacing WIP
