+++ komorebi-bar/src/cpu.rs
@@ -1,6 +1,6 @@
 use crate::config::LabelPrefix;
+use crate::render::RenderConfig;
 use crate::widget::BarWidget;
-use crate::widget::RenderConfig;
 use eframe::egui::text::LayoutJob;
 use eframe::egui::Context;
 use eframe::egui::FontId;
@@ -99,7 +99,7 @@ impl BarWidget for Cpu {
                     TextFormat::simple(font_id, ctx.style().visuals.text_color()),
                 );
 
-                config.grouping.apply_on_widget(true, config, ui, |ui| {
+                config.apply_on_widget(true, ui, |ui| {
                     if ui
                         .add(
                             Label::new(layout_job)
