+++ komorebi-bar/src/main.rs
@@ -3,11 +3,11 @@ mod battery;
 mod config;
 mod cpu;
 mod date;
-mod group;
 mod komorebi;
 mod media;
 mod memory;
 mod network;
+mod render;
 mod storage;
 mod time;
 mod ui;
@@ -29,7 +29,6 @@ use std::io::BufReader;
 use std::io::Read;
 use std::path::PathBuf;
 use std::sync::atomic::AtomicI32;
-use std::sync::atomic::AtomicU32;
 use std::sync::atomic::Ordering;
 use std::sync::Arc;
 use std::time::Duration;
@@ -44,7 +43,6 @@ use windows::Win32::UI::HiDpi::DPI_AWARENESS_CONTEXT_PER_MONITOR_AWARE_V2;
 use windows::Win32::UI::WindowsAndMessaging::EnumThreadWindows;
 use windows::Win32::UI::WindowsAndMessaging::GetWindowThreadProcessId;
 
-pub static BACKGROUND_COLOR: AtomicU32 = AtomicU32::new(0);
 pub static MAX_LABEL_WIDTH: AtomicI32 = AtomicI32::new(400);
 pub static MONITOR_LEFT: AtomicI32 = AtomicI32::new(0);
 pub static MONITOR_TOP: AtomicI32 = AtomicI32::new(0);
