+++ komorebi-bar/src/network.rs
@@ -1,6 +1,6 @@
 use crate::config::LabelPrefix;
+use crate::render::RenderConfig;
 use crate::widget::BarWidget;
-use crate::widget::RenderConfig;
 use eframe::egui::text::LayoutJob;
 use eframe::egui::Context;
 use eframe::egui::FontId;
@@ -320,7 +320,7 @@ impl BarWidget for Network {
     fn render(&mut self, ctx: &Context, ui: &mut Ui, mut config: RenderConfig) {
         if self.show_total_data_transmitted {
             for output in self.total_data_transmitted() {
-                config.grouping.apply_on_widget(true, config, ui, |ui| {
+                config.apply_on_widget(true, ui, |ui| {
                     ui.add(Label::new(output).selectable(false));
                 });
             }
@@ -328,7 +328,7 @@ impl BarWidget for Network {
 
         if self.show_network_activity {
             for output in self.network_activity() {
-                config.grouping.apply_on_widget(true, config, ui, |ui| {
+                config.apply_on_widget(true, ui, |ui| {
                     ui.add(Label::new(output).selectable(false));
                 });
             }
@@ -367,7 +367,7 @@ impl BarWidget for Network {
                     TextFormat::simple(font_id, ctx.style().visuals.text_color()),
                 );
 
-                config.grouping.apply_on_widget(true, config, ui, |ui| {
+                config.apply_on_widget(true, ui, |ui| {
                     if ui
                         .add(
                             Label::new(layout_job)
