+++ komorebi-bar/src/storage.rs
@@ -1,6 +1,6 @@
 use crate::config::LabelPrefix;
+use crate::render::RenderConfig;
 use crate::widget::BarWidget;
-use crate::widget::RenderConfig;
 use eframe::egui::text::LayoutJob;
 use eframe::egui::Context;
 use eframe::egui::FontId;
@@ -107,7 +107,7 @@ impl BarWidget for Storage {
                     TextFormat::simple(font_id.clone(), ctx.style().visuals.text_color()),
                 );
 
-                config.grouping.apply_on_widget(true, config, ui, |ui| {
+                config.apply_on_widget(true, ui, |ui| {
                     if ui
                         .add(
                             Label::new(layout_job)
